Worked on Float to int code.
Added plus sign to the return values on the stack of INT_TO_STR,
seems better than space in the sense that if the user does not
want to include a character for positive numbers, it would be
ignored anyways. Therefore a plus sign is as good as any for ignoring.