Added IO_move and IO_getxy to io.h (declaration only since these are hardware specific).
Added PICLANG_MOVE and PICLANG_GET{X,Y} functions to PICLANG to give access to IO_move and IO_getxy respectively.

Updated ed.pl to use these functions to move the cursor on the screen, using characters [hjkl], as they are used in "vi". Additionally, like vi, ed use 'i' to enter "insert" mode, ESC to leave insert mode and 'q' to leave the program (when not in insert mode.